Fix Blind Ignorance + Lancer moment
For some reason Blind Ignorance was marked as a weapon, when it shouldn't be. for some OTHER reason, this caused YOU to lose their armor in the dark world if they ever went to the light world.

For some other OTHER unrelated reason, Lancer was just not in his room for me, causing the cutscene where you get him to crash. So anyway, I fixed that.